# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Milo API",
    "version": "0.3.0"
  },
  "paths": {
...
    "/workspaces/{workspace_id}/budget": {
      "get": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Get Budget",
        "operationId": "get_budget_workspaces__workspace_id__budget_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Set Budget",
        "operationId": "set_budget_workspaces__workspace_id__budget_put",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workspaces/{workspace_id}/usage": {
      "get": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Get Usage",
        "operationId": "get_usage_workspaces__workspace_id__usage_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistant/digest": {
      "get": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Digest",
        "operationId": "digest_assistant_digest_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 36
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Id"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 1,
              "default": 24,
              "title": "Hours"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/assistant/commands": {
      "post": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Run Command",
        "operationId": "run_command_assistant_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CatchMeUp"
                  },
                  {
                    "$ref": "#/components/schemas/WriteWithMe"
                  },
                  {
                    "$ref": "#/components/schemas/TeachMe"
                  },
                  {
                    "$ref": "#/components/schemas/Pause"
                  },
                  {
                    "$ref": "#/components/schemas/Resume"
                  }
                ],
                "title": "Body",
                "discriminator": {
                  "propertyName": "command",
                  "mapping": {
                    "catch_me_up": "#/components/schemas/CatchMeUp",
                    "write_with_me": "#/components/schemas/WriteWithMe",
                    "teach_me": "#/components/schemas/TeachMe",
                    "pause": "#/components/schemas/Pause",
                    "resume": "#/components/schemas/Resume"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/budget": {
      "get": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Get Budget",
        "operationId": "get_budget_v1_workspaces__workspace_id__budget_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Set Budget",
        "operationId": "set_budget_v1_workspaces__workspace_id__budget_put",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_id}/usage": {
      "get": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Get Usage",
        "operationId": "get_usage_v1_workspaces__workspace_id__usage_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assistant/digest": {
      "get": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Digest",
        "operationId": "digest_v1_assistant_digest_get",
        "parameters": [
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          },
          {
            "name": "conversation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 36
                },
                {
                  "type": "null"
                }
              ],
              "title": "Conversation Id"
            }
          },
          {
            "name": "hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 168,
              "minimum": 1,
              "default": 24,
              "title": "Hours"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/assistant/commands": {
      "post": {
        "tags": [
          "companion",
          "budgets"
        ],
        "summary": "Run Command",
        "operationId": "run_command_v1_assistant_commands_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/CatchMeUp"
                  },
                  {
                    "$ref": "#/components/schemas/WriteWithMe"
                  },
                  {
                    "$ref": "#/components/schemas/TeachMe"
                  },
                  {
                    "$ref": "#/components/schemas/Pause"
                  },
                  {
                    "$ref": "#/components/schemas/Resume"
                  }
                ],
                "title": "Body",
                "discriminator": {
                  "propertyName": "command",
                  "mapping": {
                    "catch_me_up": "#/components/schemas/CatchMeUp",
                    "write_with_me": "#/components/schemas/WriteWithMe",
                    "teach_me": "#/components/schemas/TeachMe",
                    "pause": "#/components/schemas/Pause",
                    "resume": "#/components/schemas/Resume"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetInput": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Expected Version"
          },
          "max_actions_per_day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 1000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Actions Per Day"
          },
          "max_tokens_per_day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Tokens Per Day"
          },
          "max_cost_microusd_per_day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 2000000000.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Cost Microusd Per Day"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version"
        ],
        "title": "BudgetInput"
      },
...
      "CatchMeUp": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "command": {
            "type": "string",
            "const": "catch_me_up",
            "title": "Command"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Hours",
            "default": 24
          },
          "limit": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 20
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "workspace_id",
          "command"
        ],
        "title": "CatchMeUp"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    }
  }
}